# internal/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  uploads.CreateFolderRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
    - name
    type: object
  uploads.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  uploads.FileDetailResponse:
    properties:
      created_at:
        type: string
      extension:
        type: string
      file_type:
        type: string
      folder:
        $ref: '#/definitions/uploads.FolderResponse'
      id:
        type: integer
      is_public:
        type: boolean
      mime_type:
        type: string
      name:
        type: string
      original_name:
        type: string
      size:
        type: integer
      storage_path:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: integer
    type: object
  uploads.FileResponse:
    properties:
      created_at:
        type: string
      extension:
        type: string
      file_type:
        type: string
      folder_name:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      mime_type:
        type: string
      name:
        type: string
      original_name:
        type: string
      size:
        type: integer
      storage_path:
        type: string
      updated_at:
        type: string
    type: object
  uploads.FolderContentsResponse:
    properties:
      file_total:
        type: integer
      files:
        items:
          $ref: '#/definitions/uploads.FileResponse'
        type: array
      folder_total:
        type: integer
      folders:
        items:
          $ref: '#/definitions/uploads.FolderResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
    type: object
  uploads.FolderResponse:
    properties:
      created_at:
        type: string
      file_count:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      storage_path:
        type: string
      updated_at:
        type: string
    type: object
  uploads.FolderResponseWithChildren:
    properties:
      children:
        items:
          $ref: '#/definitions/uploads.FolderResponseWithChildren'
        type: array
      created_at:
        type: string
      file_count:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      storage_path:
        type: string
      updated_at:
        type: string
    type: object
  uploads.ListResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  uploads.MessageResponse:
    properties:
      message:
        type: string
    type: object
  uploads.MoveFileRequest:
    properties:
      folder_id:
        type: integer
    required:
    - folder_id
    type: object
  uploads.UpdateFolderRequest:
    properties:
      name:
        type: string
    type: object
  uploads.UploadResponse:
    properties:
      failed:
        items:
          type: string
        type: array
      files:
        items:
          $ref: '#/definitions/uploads.FileResponse'
        type: array
    type: object
...
info:
  contact: {}
  description: Backend API for ASSMI Super Shop ERP.
  title: ASSMI Super Shop ERP API
  version: "1.0"
paths:
...
  /uploads/files:
    get:
      description: Get all files uploaded by the user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.ListResponse'
      security:
      - BearerAuth: []
      summary: Get user files
      tags:
      - Uploads----------Files
    post:
      consumes:
      - multipart/form-data
      description: Upload a single file to a specific folder or root
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Folder ID
        in: formData
        name: folder_id
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/uploads.FileResponse'
      security:
      - BearerAuth: []
      summary: Upload a single file
      tags:
      - Uploads----------Files
  /uploads/files/{file_id}:
    delete:
      description: Soft delete a file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - Uploads----------Files
    get:
      description: Get detailed information about a specific file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.FileDetailResponse'
      security:
      - BearerAuth: []
      summary: Get file details
      tags:
      - Uploads----------Files
  /uploads/files/{file_id}/download:
    get:
      description: Download a file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Download file
      tags:
      - Uploads----------Files
  /uploads/files/{file_id}/move:
    post:
      consumes:
      - application/json
      description: Move a file to a different folder
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: integer
      - description: Destination folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/uploads.MoveFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.FileResponse'
      security:
      - BearerAuth: []
      summary: Move file to another folder
      tags:
      - Uploads----------Files
  /uploads/files/{file_id}/restore:
    post:
      description: Restore a deleted file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.MessageResponse'
      security:
      - BearerAuth: []
      summary: Restore file
      tags:
      - Uploads----------Files
...
  /uploads/files/batch:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple files at once
      parameters:
      - collectionFormat: csv
        description: Files to upload
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      - description: Folder ID
        in: formData
        name: folder_id
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/uploads.UploadResponse'
      security:
      - BearerAuth: []
      summary: Upload multiple files
      tags:
      - Uploads----------Files
...
  /uploads/folders:
    get:
      description: Get all root folders (folders without parent) for the user with
        their child folders nested
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/uploads.FolderResponseWithChildren'
            type: array
      security:
      - BearerAuth: []
      summary: Get root folders with subfolders
      tags:
      - Uploads----------Folders
    post:
      consumes:
      - application/json
      description: Create a new folder in the specified parent or root
      parameters:
      - description: Folder details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/uploads.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/uploads.FolderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/uploads.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new folder No need parent id if you want to create root folder
      tags:
      - Uploads----------Folders
  /uploads/folders/{folder_id}:
    delete:
      description: Soft delete a folder and its contents
      parameters:
      - description: Folder ID
        in: path
        name: folder_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - Uploads----------Folders
    patch:
      consumes:
      - application/json
      description: Update folder details
      parameters:
      - description: Folder ID
        in: path
        name: folder_id
        required: true
        type: integer
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/uploads.UpdateFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.FolderResponse'
      security:
      - BearerAuth: []
      summary: Update folder
      tags:
      - Uploads----------Folders
  /uploads/folders/{folder_id}/contents:
    get:
      description: Get subfolders and files in a folder with pagination. If no folder_id
        is provided, returns root folder contents.
      parameters:
      - description: Folder ID (optional, defaults to root)
        in: path
        name: folder_id
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.FolderContentsResponse'
      security:
      - BearerAuth: []
      summary: Get folder contents like files
      tags:
      - Uploads----------Folders
  /uploads/folders/{folder_id}/restore:
    post:
      description: Restore a deleted folder
      parameters:
      - description: Folder ID
        in: path
        name: folder_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.MessageResponse'
      security:
      - BearerAuth: []
      summary: Restore folder
      tags:
      - Uploads----------Folders
  /uploads/folders/contents:
    get:
      description: Get subfolders and files in a folder with pagination. If no folder_id
        is provided, returns root folder contents.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploads.FolderContentsResponse'
      security:
      - BearerAuth: []
      summary: Get folder contents like files
      tags:
      - Uploads----------Folders
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"